backend-python-agent/requirements.txt
fastapi
uvicorn[standard]
httpx
requests
grpcio
grpcio-tools

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi

